Refactorization-> moved models in a folder.
