+++ Sdl.Community.Toolkit.LanguagePlatform/Models/Xliff/Settings.cs
@@ -2,7 +2,7 @@
 using Sdl.LanguagePlatform.Core.Tokenization;
 using Sdl.LanguagePlatform.TranslationMemory;
 
-namespace Sdl.Community.Toolkit.LanguagePlatform.Models
+namespace Sdl.Community.Toolkit.LanguagePlatform.Models.Xliff
 {
 	/// <summary>
 	/// Settings
