+++ Sdl.Community.Toolkit.LanguagePlatform/Models/Xliff/TargetTranslation.cs
@@ -1,8 +1,9 @@
 ﻿using System.Globalization;
 using System.Xml.Serialization;
+using Sdl.Community.Toolkit.LanguagePlatform.XliffConverter;
 using Sdl.LanguagePlatform.Core;
 
-namespace Sdl.Community.Toolkit.LanguagePlatform.XliffConverter
+namespace Sdl.Community.Toolkit.LanguagePlatform.Models.Xliff
 {
 	public class TargetTranslation
 	{
