+++ Sdl.Community.Toolkit.LanguagePlatform/Models/Xliff/TranslationOption.cs
@@ -1,6 +1,6 @@
 ﻿using System.Xml.Serialization;
 
-namespace Sdl.Community.Toolkit.LanguagePlatform.XliffConverter
+namespace Sdl.Community.Toolkit.LanguagePlatform.Models.Xliff
 {
 	public class TranslationOption
 	{
